+++ src/main/java/subway/line/domain/Line.java
@@ -1,4 +1,6 @@
-package subway.line;
+package subway.line.domain;
+
+import subway.line.presentation.LineRequest;
 
 public class Line {
