+++ src/main/java/subway/section/domain/SectionDao.java
@@ -1,10 +1,11 @@
-package subway.section;
+package subway.section.domain;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import subway.section.domain.Section;
 
 import java.util.List;
