+++ src/main/java/subway/section/domain/strategy/UpSectionGenerateStrategy.java
@@ -1,7 +1,7 @@
-package subway.section.strategy;
+package subway.section.domain.strategy;
 
 import subway.exceptions.sectionExceptions.SectionIllegalDistanceException;
-import subway.section.Section;
+import subway.section.domain.Section;
 
 public class UpSectionGenerateStrategy implements SectionGenerateStrategy {
     private Section downSection;
