+++ src/test/java/subway/section/SectionTest.java
@@ -4,6 +4,7 @@ package subway.section;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import subway.section.domain.Section;
 
 import static org.assertj.core.api.Assertions.assertThat;
